+++ fabric/src/main/java/com/faboslav/structurify/fabric/modcompat/TerraCompat.java
@@ -1,0 +1,56 @@
+package com.faboslav.structurify.fabric.modcompat;
+
+import com.faboslav.structurify.common.modcompat.ModCompat;
+import net.minecraft.resources.ResourceLocation;
+
+import java.lang.reflect.Method;
+import java.util.ArrayList;
+import java.util.EnumSet;
+import java.util.List;
+import java.util.Map;
+
+public final class TerraCompat implements ModCompat
+{
+	@Override
+	public EnumSet<Type> compatTypes() {
+		return EnumSet.of(Type.BIOME_REPLACER);
+	}
+
+	@Override
+	public List<String> getReplacedBiomes(List<String> biomes) {
+		var terraBiomes = new ArrayList<String>();
+		var terraBiomeMap = this.getTerraBiomes("com.dfsek.terra.mod.util.BiomeUtil", "getTerraBiomeMap");
+
+		if(terraBiomeMap == null) {
+			terraBiomeMap = this.getTerraBiomes("com.dfsek.terra.mod.util.MinecraftUtil", "getTerraBiomeMap");
+		}
+
+		if(terraBiomeMap != null) {
+			terraBiomeMap.forEach((originalBiome, replacementBiomes) -> {
+				var vanillaBiomeName = originalBiome.toString();
+
+				if (biomes.contains(vanillaBiomeName)) {
+					for (var replacementBiome : replacementBiomes) {
+						terraBiomes.add(replacementBiome.toString());
+					}
+				}
+			});
+		}
+
+		return terraBiomes;
+	}
+
+	public Map<ResourceLocation, List<ResourceLocation>> getTerraBiomes(String className, String methodName) {
+		try {
+			Class<?> clazz = Class.forName(className);
+			Method method = clazz.getDeclaredMethod(methodName);
+			method.setAccessible(true);
+			Object result = method.invoke(null);
+			return (Map<ResourceLocation, List<ResourceLocation>>) result;
+
+		} catch (Exception e) {
+			e.printStackTrace();
+			return null;
+		}
+	}
+}